Move `CacheHttp` out of `http` module into `model`
