+++ src/builder/create_interaction_response.rs
@@ -13,8 +13,7 @@ use super::{
     EditAttachments,
 };
 #[cfg(feature = "http")]
-use crate::http::{CacheHttp, Http};
-use crate::internal::prelude::*;
+use crate::http::Http;
 use crate::model::prelude::*;
 
 /// [Discord docs](https://docs.discord.com/developers/interactions/receiving-and-responding#interaction-response-object).
