+++ src/builder/create_interaction_response_followup.rs
@@ -9,10 +9,6 @@ use super::{
     CreatePoll,
     EditAttachments,
 };
-#[cfg(feature = "http")]
-use crate::http::CacheHttp;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// [Discord docs](https://docs.discord.com/developers/interactions/receiving-and-responding#create-followup-message)
