+++ src/builder/create_soundboard.rs
@@ -1,10 +1,6 @@
 use std::borrow::Cow;
 
 use super::DataUri;
-#[cfg(feature = "http")]
-use crate::http::CacheHttp;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// A builder to create a soundboard sound.
