+++ src/builder/create_sticker.rs
@@ -4,8 +4,6 @@ use super::CreateAttachment;
 #[cfg(feature = "http")]
 use crate::http::Http;
 #[cfg(feature = "http")]
-use crate::internal::prelude::*;
-#[cfg(feature = "http")]
 use crate::model::prelude::*;
 
 /// A builder to create a guild sticker
