+++ src/builder/create_thread.rs
@@ -4,8 +4,6 @@ use nonmax::NonMaxU16;
 
 #[cfg(feature = "http")]
 use crate::http::Http;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// Discord docs:
