+++ src/builder/edit_guild_incident_actions.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "http")]
 use crate::http::Http;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// A builder for editing guild incident actions.
