+++ src/builder/edit_guild_widget.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "http")]
 use crate::http::Http;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// A builder to specify the fields to edit in a [`GuildWidget`].
