+++ src/builder/edit_message.rs
@@ -7,10 +7,6 @@ use super::{
     CreateEmbed,
     EditAttachments,
 };
-#[cfg(feature = "http")]
-use crate::http::CacheHttp;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// A builder to specify the fields to edit in an existing message.
