+++ src/builder/edit_voice_state.rs
@@ -1,7 +1,5 @@
 #[cfg(feature = "http")]
 use crate::http::Http;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// A builder which edits a user's voice state, to be used in conjunction with
