+++ src/builder/edit_webhook.rs
@@ -3,8 +3,6 @@ use std::borrow::Cow;
 use super::DataUri;
 #[cfg(feature = "http")]
 use crate::http::Http;
-#[cfg(feature = "http")]
-use crate::internal::prelude::*;
 use crate::model::prelude::*;
 
 /// [Discord docs](https://docs.discord.com/developers/resources/webhook#modify-webhook)
