+++ src/builder/edit_webhook_message.rs
@@ -8,8 +8,6 @@ use super::{
     EditAttachments,
 };
 #[cfg(feature = "http")]
-use crate::http::CacheHttp;
-#[cfg(feature = "http")]
 use crate::internal::prelude::*;
 use crate::model::prelude::*;
