+++ src/model/application/command_interaction.rs
@@ -13,7 +13,7 @@ use crate::builder::{
     EditInteractionResponse,
 };
 #[cfg(feature = "model")]
-use crate::http::{CacheHttp, Http};
+use crate::http::Http;
 use crate::model::prelude::*;
 
 /// An interaction when a user invokes a slash command.
